requirements.txt
streamlit==1.24.1
google-genai>=1.0.0
pandas
numpy<=1.25.1
